count songCount and albumCount by artist

+++ qm-music-parent/qm-music-core/src/main/java/com/github/chenqimiao/service/UserStarService.java
@@ -23,7 +23,7 @@ public interface UserStarService {
      Long starredTime(StarInfoRequest starInfoRequest);
 
 
-     Map<Integer, Long> batchQueryStarredTime(BatchStarInfoRequest batchStarInfoRequest);
+     Map<Long, Long> batchQueryStarredTime(BatchStarInfoRequest batchStarInfoRequest);
 
 
      List<UserStarDTO> queryUserStarByUserId(Long userId);
